Move direction and credential values into detector_payload.py; changelog
Drift Bot reported five findings on this PR that the code disproves (line
numbers posted on the PR): it reads roughly the first 500 lines of each
changed file plus main's copy, and the new code sat at detector_surface.py
lines 313-453 while the NormStep shape comment it trusts still listed none
of the new step fields.

- New leaf module clawmetry/detector_payload.py holds the direction
  classification (_write_hosts, _segment_write_hosts) and the credential value
  table (_secret_value_categories, SECRET_VALUE_OWNERS, host_owned_by), moved
  byte-for-byte. detector_surface.py is back to main's exact bytes.
- detectors.py: NormStep comment lists paths/cmd/hosts/write_hosts/
  secret_values; session_profile docstring names write_hosts.
- CHANGELOG entry under Unreleased.

No behaviour change: 414 passed, 10 skipped across the detector, Guard,
policy and schema suites.
