+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Unreleased
 
+### Added: Guard sees write direction, credential values and remote privilege (2026-09-11)
+- **Why:** the Hugging Face swarm scorecard (clawmetry.com/blog/could-clawmetry-have-caught-the-hugging-face-swarm) found five things a single swarm member already showed that the behavioural detectors could not see: WebDAV `MKCOL`/`PUT` to an Artifactory host every session was allowed to reach (row 2), an upload to huggingface.co (row 10), Hugging Face write tokens read from output and posted to a board (row 9), admin minted through a token endpoint (row 3), and cluster-admin (row 12). `network_egress` keyed on host only, so a read and a write looked identical; `credential_access` matched secret locations only (0 of 45 values in the vendored corpus); `privilege_change` knew the local OS only.
+- **Direction:** each shell segment is classified as a read or a write (curl method or body flags, matched case-sensitively so `-x` and `-f` stay reads and `-G` stays a GET; wget `--post-*`; httpie; `twine`/`npm`/`hf`/`cargo`/`gem` publish; `s3`/`gs` copies and scp/rsync by destination; HTTP tools with a `method` argument). An upload that names no URL is still egress. `network_egress` gains the `write_to_read_only_host` ground: a write to a host the cohort has only ever read from. `guard_egress_hosts` gains `writes`, `write_first_seen` and `dir_since`.
+- **Two clocks keep it honest:** a host is read-only only once its direction has been watched for the settle window (existing rows migrate with `dir_since` empty, so an upgrade cannot flag every host), and a first write inside the window does not count, so a swarm's first writer cannot vouch for the rest.
+- **Credential values:** token shapes (Hugging Face, GitHub, AWS, Google, OpenAI and Anthropic, Slack, Stripe, JWT, PEM, and the pipelock-community DLP rules ported with a left boundary) in call arguments, heredocs included, and in tool output. Only the category is recorded. Sent to a host that does not own that kind of token, or found in output and then sent anywhere: critical. Placeholders such as `AKIA...EXAMPLE` and kebab-case slugs are rejected.
+- **Remote privilege:** role bindings, `system:admin`/`masters` impersonation, AWS IAM key and policy grants, GCP and Azure role grants, privileged or host-root-mounted containers, and writing calls to Artifactory token and plugin and Jenkins script-console endpoints. All critical; reading or listing the same resources stays quiet.
+- **Verified:** `tests/test_detectors_swarm_gaps.py` (62 tests); the pipelock corpus value lane moved from 0/45 to 45/45 with 0 near-miss over-fires (`VALUE_LANE_BASELINE` updated deliberately); 0 findings from any new lane across 400 real sessions (13,434 steps) on a development machine.
+- **Not claimed:** writes or credentials a program handles internally. These read tool arguments and tool output, and every finding says so.
+
 ### Fixed: the stall detector called every sleeping laptop broken (2026-09-11)
 - **Why:** the daemon stall reporter shipped in 0.12.855 immediately produced 8 distinct installs and 9 events in under a day, across Darwin and Linux and five Python versions — against the one bootstrap signature that has 1 install in 30 days. Six issues were auto-filed overnight (#5829–#5834). The volume was the tell: the reports were not trustworthy.
 - **The bug is in the clock.** `last_sync` is a **wall-clock** stamp and the check was `now - last_sync > 3600`. Wall-clock elapsed time cannot distinguish "the ingest loop is wedged" from "this machine was not running", so two entirely healthy cases reported as broken: a laptop suspended overnight wakes with an age of hours while the next cycle completes seconds later (the watchdog ticks within 30s of wake, before it lands); and a daemon that just started on a machine powered off overnight reads yesterday's `last_sync`. Every node auto-updates from PyPI, so **every daemon restarted when 0.12.855 shipped** — which is when the reports began. 6 of the 8 installs were Darwin. Laptops sleep.
