+++ docs/MODULE_MAP.md
@@ -4,7 +4,7 @@
 > `python3 scripts/gen_module_map.py` (CI fails on drift via
 > `tests/test_module_map_drift.py`).
 
-237 modules, 81 Flask blueprints. `CLAUDE.md` carries a short curated table of the ones you reach for most often; this is the whole list.
+238 modules, 81 Flask blueprints. `CLAUDE.md` carries a short curated table of the ones you reach for most often; this is the whole list.
 
 Size bands are deliberately coarse so this file does not churn on every PR: **small** is under 200 lines, **medium** under 1k, **large** under 5k, **huge** is 5k and up.
 
@@ -146,6 +146,7 @@ The pip-installable package: CLI, sync daemon, DuckDB store, detectors, enforcem
 | `clawmetry/detector_behaviour.py` | medium | Is this agent doing something it does not normally do? |
 | `clawmetry/detector_calibration.py` | medium | How a detector decides what "too many" means, for THIS runtime and THIS team. |
 | `clawmetry/detector_money.py` | small | What a finding costs, and therefore what to look at first. |
+| `clawmetry/detector_payload.py` | medium | What a tool call SENT, and what it CARRIED. |
 | `clawmetry/detector_surface.py` | medium | What a tool call actually touched, and what a finding may repeat back. |
 | `clawmetry/detectors.py` | large | research-backed, judge-free, CPU-cheap trajectory anomaly detectors over a session's recent event sequence (issue #2999). |
 | `clawmetry/deterministic_evaluators.py` | medium | cheap, code-based checks on sessions. |
